requirements.txt
jupyter
python-dotenv
pydantic


# Environment specific..
# tensorflow

# Note: this requires the rust tool chain, m1 setup: https://sourabhbajaj.com/mac-setup/Rust/
transformers
datasets >= 1.8.0
sentencepiece != 0.1.92


Django>=4.1.2
django-autocomplete-light>=3.9.4
djhacker>=0.2.3
fastapi~=0.81
pydantic~=1.10
uvicorn[standard]~=0.18
starlette

